Add files via upload
TO USE : in the command line terminal, first type the size of rows, columns and optionally, the seed for for the random factor.

First iteration that works with already set positions. Further updates will include the ability to choose the start/end pos and to animate the path. 

Part of VU Bachelor's CS course. (Vrije Universiteit Amsterdam)